// src/types.ts
export type Data = {
  name: string;
  class: string;
  points: {
    "1": number | string;
    "2": number | string;
    "3": number | string;
    "4": number | string;
    "6": number | string;
    "7": number | string;
    "9 - 13": number | string;
    "Test kiến thức": number | string;
    "Final project": number | string;
    Tổng: number | string;
  };
}[];